requirements.txt
pytest
pymssql
python-dotenv
# python3 -m venv .venv
# source .venv/bin/activate
# pip3 install -r requirements.txt